[refactor] change test shell file name
