+++ tests/test_shell.py
@@ -1,250 +1,72 @@
 import tempfile
-
 import pytest
 from pytest_mock import MockerFixture
 from shell import *
 
-
 @pytest.fixture
 def shell_app(mocker):
-    ssd_driver_mock = mocker.Mock(spec=SSDDriver)
-    test_shell_app = TestShellApp(ssd_driver_mock)
+    ssd_driver = SSDDriver()
+    test_shell_app = TestShellApp(ssd_driver)
 
     return test_shell_app
 
-
-def test_shell_write_subprocess(shell_app):
+@pytest.mark.parametrize("wrong_cmd", ["exi", "rea", "wri", "hel", " ", "0_", "*"])
+def test_shell_wrong_cmd(shell_app, mocker: MockerFixture, wrong_cmd, capsys):
     # Arrange
-    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
+    cmd = [f"{wrong_cmd}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
     # Act
-    ret = shell_app.write(address="0", value="0x00000000")
-
-    # Assert
-    assert ret == WRITE_SUCCESS
-    shell_app._ssd_driver.run_ssd_write.assert_called_once_with(address="0", value="0x00000000")
-
-
-@pytest.mark.parametrize("wrong_address", ["-1", "100", "hello", "0.5", "-0.5", "123", ";"])
-def test_shell_write_wrong_address(shell_app, wrong_address):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
-
-    # Act
-    ret = shell_app.write(address=wrong_address, value="0x00000000")
-
-    # Assert
-    assert ret == WRITE_ERROR
-
-
-@pytest.mark.parametrize("wrong_value", ["AA", "0xHELLO", "ox11", "0xaaaaaaaaaa", "-0xa", "1234", ";"])
-def test_shell_write_wrong_value(shell_app, wrong_value):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
-
-    # Act
-    ret = shell_app.write(address="0", value=wrong_value)
-
-    # Assert
-    assert ret == WRITE_ERROR
-
-
-@pytest.mark.parametrize("valid_address", [str(x) for x in range(100)])
-def test_shell_write_test_valid_address(shell_app, valid_address):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
-
-    # Act
-    ret = shell_app.write(address=valid_address, value="0x00000001")
-
-    # Assert
-    assert ret == WRITE_SUCCESS
-    shell_app._ssd_driver.run_ssd_write.assert_called_once_with(address=valid_address, value="0x00000001")
-
-
-@pytest.mark.parametrize("valid_value", ["0xa", "0xab", "0xabc", "0xabcd", "0xabcde", "0xabcdef", "0xabcdeff"])
-def test_shell_write_valid_value(shell_app, valid_value):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
-
-    # Act
-    ret = shell_app.write(address="0", value=valid_value)
-
-    # Assert
-    assert ret == WRITE_SUCCESS
-
-
-def test_shell_read(shell_app, capsys):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_read.return_value = 0
-    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000000"
-
-    # Act
-    ret = shell_app.read(address="0")
+    shell_app.run_shell(cmd_len)
     captured = capsys.readouterr()
 
     # Assert
-    assert ret == READ_SUCCESS
-    assert '[Read] LBA 00 : 0x00000000' in captured.out
-    shell_app._ssd_driver.run_ssd_read.assert_called_once_with(address="0")
-
-
-def test_shell_read_with_real(capsys):
-    # Arrange
-    shell_app = TestShellApp()
-
-    # Act
-    ret = shell_app.read(address="0")
-    captured = capsys.readouterr()
-
-    # Assert
-    assert ret == READ_SUCCESS
-
-
-def test_shell_read_after_write(shell_app, capsys):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
-    shell_app._ssd_driver.run_ssd_read.return_value = READ_SUCCESS
-    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000000"
-
-    # Act
-    ret_write = shell_app.write(address="0", value="0x00000000")
-    ret_read = shell_app.read(address="0")
-    captured = capsys.readouterr()
-
-    # Assert
-    assert ret_write == WRITE_SUCCESS
-    assert ret_read == READ_SUCCESS
-    assert '[Read] LBA 00 : 0x00000000' in captured.out
-
-    shell_app._ssd_driver.run_ssd_write.assert_called_once_with(address="0", value="0x00000000")
-    shell_app._ssd_driver.run_ssd_read.assert_called_once_with(address="0")
-
-
-def test_shell_read_after_read(shell_app, capsys):
-    # Arrange & Act
-    shell_app._ssd_driver.run_ssd_read.return_value = READ_SUCCESS
-    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000000"
-    ret0 = shell_app.read(address="0")
-
-    shell_app._ssd_driver.run_ssd_read.return_value = READ_SUCCESS
-    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000010"
-    ret1 = shell_app.read(address="1")
-
-    captured = capsys.readouterr()
-
-    # Assert
-    assert ret0 == READ_SUCCESS
-    assert ret1 == READ_SUCCESS
-    assert '[Read] LBA 00 : 0x00000000' in captured.out
-    assert '[Read] LBA 01 : 0x00000010' in captured.out
-
-
-def test_shell_cmd_exit_success(shell_app, mocker: MockerFixture):
-    mocker.patch("builtins.input", side_effect=["exit"])
-    mock_method = mocker.patch("shell.TestShellApp.exit")
-    mock_method.return_value = 0
-    shell_app.run_shell(1)
-
-    mock_method.assert_called_once()
-
-
-def test_shell_cmd_read_success(shell_app, mocker: MockerFixture):
-    mocker.patch("builtins.input", side_effect=["read 1"])
-    mock_method = mocker.patch("shell.TestShellApp.read")
-    mock_method.return_value = 0
-    shell_app.run_shell(1)
-
-    mock_method.assert_called_once()
-
-
-def test_shell_cmd_write_success(shell_app, mocker: MockerFixture):
-    mocker.patch("builtins.input", side_effect=["write 0 0x00000001"])
-    mock_method = mocker.patch("shell.TestShellApp.write")
-    mock_method.return_value = 0
-    shell_app.run_shell(1)
-
-    mock_method.assert_called_once()
-
-
-def test_shell_cmd_help_success(shell_app, mocker: MockerFixture):
-    mocker.patch("builtins.input", side_effect=["help"])
-    mock_method = mocker.patch("shell.TestShellApp.help")
-    mock_method.return_value = 0
-    shell_app.run_shell(1)
-
-    mock_method.assert_called_once()
-
-
-def test_shell_cmd_fullread_success(shell_app, mocker: MockerFixture):
-    mocker.patch("builtins.input", side_effect=["fullread"])
-    mock_method = mocker.patch("shell.TestShellApp.full_read")
-    mock_method.return_value = 0
-    shell_app.run_shell(1)
-
-    mock_method.assert_called_once()
-
-
-def test_shell_cmd_fullwrite_success(shell_app, mocker: MockerFixture):
-    mocker.patch("builtins.input", side_effect=["fullwrite 0x00010000"])
-    mock_method = mocker.patch("shell.TestShellApp.full_write")
-    mock_method.return_value = 0
-    shell_app.run_shell(1)
-
-    mock_method.assert_called_once()
-
-
-@pytest.mark.parametrize("input", ["exi", "rea", "wri", "hel"])
-def test_shell_wrong_cmd(shell_app, mocker: MockerFixture, input, capsys):
-    mocker.patch("builtins.input", side_effect=[input])
-
-    shell_app.run_shell(1)
-    captured = capsys.readouterr()
-
     assert 'INVALID COMMAND' in captured.out
 
+@pytest.mark.parametrize("wrong_cmd_args", ["write", "read 1 2", "fullwrite", "fullread 0000", "exit 3", "help -", \
+                                            "1_ 02", "2_ s", "2_ a", "4_ *"])
+def test_shell_wrong_cmd_args(shell_app, mocker: MockerFixture, wrong_cmd_args, capsys):
+    # Arrange
+    cmd = [f"{wrong_cmd_args}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
-def test_shell_wrong_cmd_empty(shell_app, mocker: MockerFixture, capsys):
-    mocker.patch("builtins.input", side_effect=[""])
-
-    shell_app.run_shell(1)
+    # Act
+    shell_app.run_shell(cmd_len)
     captured = capsys.readouterr()
 
+    # Assert
     assert 'INVALID COMMAND' in captured.out
 
+def test_shell_cmd_exit_success(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["exit"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
-@pytest.mark.parametrize("input", ["read 1 2", "write", "fullwrite", "exit 3"])
-def test_shell_wrong_cmd_args(shell_app, mocker: MockerFixture, input, capsys):
-    mocker.patch("builtins.input", side_effect=["read 1 2"])
-
-    shell_app.run_shell(1)
-    captured = capsys.readouterr()
-
-    assert 'INVALID COMMAND' in captured.out
-
-
-def test_shell_subprocess_cmd():
-    pass
-
-
-def test_shell_exit(shell_app, mocker: MockerFixture):
-    mocker.patch("builtins.input", side_effect=["exit"])
-    exit_mock = mocker.patch("shell.TestShellApp.exit", side_effect=SystemExit(0))
-
+    # Act
     with pytest.raises(SystemExit) as e:
-        shell_app.run_shell(1)
-
-    exit_mock.assert_called_once()
-    assert e.value.code == 0
-
-
-def test_shell_help(capsys):
-    test_shell_app = TestShellApp()
-    test_shell_app.help()
+        shell_app.run_shell(cmd_len)
 
     captured = capsys.readouterr()
-    output_lines = captured.out.strip().splitlines()
 
+    # Assert
+    assert e.value.code == 0
+    assert 'INVALID COMMAND' not in captured.out
+
+def test_shell_cmd_help_success(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["help"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' not in captured.out
     expected_lines = [
         "팀명: BestReviewer",
         "팀장: 이장희 / 팀원: 김대용, 최도현, 박윤상, 최동희, 안효민, 김동훈",
@@ -263,87 +85,444 @@ def test_shell_help(capsys):
         "  help                               : 도움말 출력",
         "  exit                               : 종료"
     ]
-
     for line in expected_lines:
-        assert line in output_lines
+        assert line in captured.out
 
-
-def test_shell_full_read(shell_app, mocker: MockerFixture, capsys):
+def test_shell_write(shell_app, mocker: MockerFixture, capsys):
     # Arrange
-    shell_app._ssd_driver.run_ssd_read.side_effect = [0] * 100
-    shell_app._ssd_driver.get_ssd_output.side_effect = ["0x00000000"] * 100
+    cmd = ["write 0 0x00000001"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
     # Act
-    ret = shell_app.full_read()
+    ret = shell_app.run_shell(cmd_len)
     captured = capsys.readouterr()
-    assert ret == READ_SUCCESS
-    assert shell_app._ssd_driver.run_ssd_read.call_count == 100
-
-
-def test_shell_full_write(shell_app, mocker: MockerFixture):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_write.side_effect = [WRITE_SUCCESS] * 100
-
-    # Act
-    ret = shell_app.full_write(value="0x12345678")
 
     # Assert
-    assert ret == WRITE_SUCCESS
-    assert shell_app._ssd_driver.run_ssd_write.call_count == 100
+    assert '[Write] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
 
+@pytest.mark.parametrize("valid_address", [str(x) for x in range(100)])
+def test_shell_write_test_valid_address(shell_app, valid_address, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write {valid_address} 0x00000001"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
-def test_shell_full_write_wrong_format():
-    pass
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
 
+    # Assert
+    assert '[Write] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("valid_value", ["0xa", "0xab", "0xabc", "0xabcd", "0xabcde", "0xabcdef", "0xabcdeff", "0x00000000000000000001"])
+def test_shell_write_valid_value(shell_app, valid_value, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write 00 {valid_value}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Write] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("wrong_address", ["-1", "100", "hello", "0.5", "-0.5", "123", ";", " ", "0x00"])
+def test_shell_write_wrong_address(shell_app, wrong_address, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write {wrong_address} 0x12345678"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Write] Done' not in captured.out
+    assert 'INVALID COMMAND' in captured.out
+
+@pytest.mark.parametrize("wrong_value", ["AA", "0xHELLO", "ox11", "0xaaaaaaaaaa", "-0xa", "1234", ";", " ", "0xA00000000000000001"])
+def test_shell_write_wrong_value(shell_app, wrong_value, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write 00 {wrong_value}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Write] Done' not in captured.out
+    assert 'INVALID COMMAND' in captured.out
+
+def test_shell_read(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["write 00 0x00000000", "read 0"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Write] Done' in captured.out
+    assert '[Read] LBA 00 : 0x00000000' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("valid_address", [str(x) for x in range(100)])
+def test_shell_read_test_valid_address(shell_app, valid_address, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write {valid_address} 0x00000001", f"read {valid_address}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Write] Done' in captured.out
+    assert f'[Read] LBA {int(valid_address):02} : 0x00000001' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("valid_value", ["0xa", "0xab", "0xabc", "0xabcd", "0xabcde", "0xabcdef", "0xabcdeff", "0x00000000000000000001"])
+def test_shell_read_valid_value(shell_app, valid_value, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write 00 {valid_value}", f"read 00"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    int_value = int(valid_value, 16)
+    assert '[Write] Done' in captured.out
+    assert f'[Read] LBA 00 : 0x{int_value:08X}' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("wrong_address", ["-1", "100", "hello", "0.5", "-0.5", "123", ";", " ", "0x00"])
+def test_shell_read_wrong_address(shell_app, wrong_address, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write {wrong_address} 0x12345678", f"read {wrong_address}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Write] Done' not in captured.out
+    assert 'INVALID COMMAND' in captured.out
+
+@pytest.mark.parametrize("wrong_value", ["AA", "0xHELLO", "ox11", "0xaaaaaaaaaa", "-0xa", "1234", ";", " ", "0xA00000000000000001"])
+def test_shell_read_after_write_wrong_value(shell_app, wrong_value, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"write 00 {wrong_value}", f"read 00"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Write] Done' not in captured.out
+    assert 'INVALID COMMAND' in captured.out
+
+def test_shell_erase(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["erase 0 1"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("valid_address", [str(x) for x in range(100)])
+def test_shell_erase_with_valid_address(shell_app, valid_address, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"erase {valid_address} 1"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("wrong_address", ["-1", "100", "hello", "0.5", "-0.5", "123", ";", " ", "0x00"])
+def test_shell_erase_with_wrong_address(shell_app, wrong_address, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"erase {wrong_address} 1"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase] Done' not in captured.out
+    assert 'INVALID COMMAND' in captured.out
+
+@pytest.mark.skip
+@pytest.mark.parametrize("valid_address", [str(x) for x in range(100)])
+@pytest.mark.parametrize("valid_size", ["1", "100", "4294967295", "0", "-1", "-100", "-4294967295"])
+def test_shell_erase_with_valid_size(shell_app, valid_address, valid_size, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"erase {valid_address} {valid_size}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("invalid_size", ["0.5", "-1.5", "0xa", "-0x1", "0xF", "-", " ", "null", "A"])
+def test_shell_erase_with_invalid_size(shell_app, invalid_size, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"erase 00 {invalid_size}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase] Done' not in captured.out
+    assert 'INVALID COMMAND' in captured.out
+
+def test_shell_erase_range(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["erase_range 0 1", "erase_range 1 0"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase_range] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("valid_range", [("0", "0"), ("99", "99"), ("0", "99"), ("99", "0")])
+def test_shell_erase_range_with_valid_range(shell_app, valid_range, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"erase_range {valid_range[0]} {valid_range[1]}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase_range] Done' in captured.out
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("invalid_range", [("-1", "10"), ("1", "100"), ("1.5", "10"), ("10", "10.5")
+                                           , ("a", "10.5"), ("10", ".5")])
+def test_shell_erase_range_invalid_range(shell_app, invalid_range, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"erase_range {invalid_range[0]} {invalid_range[1]}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert '[Erase_range] Done' not in captured.out
+    assert 'INVALID COMMAND' in captured.out
+
+def test_shell_cmd_fullwrite(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["fullwrite 0x00000001"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' not in captured.out
+
+def test_shell_cmd_fullread(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["fullwrite 0x00000001", "fullread"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' not in captured.out
+    for address in range(100):
+        assert f'[Read] LBA {address:02} : 0x00000001' in captured.out
+
+@pytest.mark.skip
+@pytest.mark.parametrize("valid_value", ["0xa", "0xab", "0xabc", "0xabcd", "0xabcde", "0xabcdef", "0xabcdeff", "0x00000000000000000001"])
+def test_shell_cmd_fullwrite_with_valid_value(shell_app, valid_value, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"fullwrite {valid_value}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' not in captured.out
+
+@pytest.mark.parametrize("wrong_value", ["AA", "0xHELLO", "ox11", "0xaaaaaaaaaa", "-0xa", "1234", ";", " ", "0xA00000000000000001"])
+def test_shell_cmd_fullwrite_with_wrong_value(shell_app, wrong_value, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"fullwrite {wrong_value}"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' in captured.out
+
+def test_shell_cmd_fullread(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = ["fullwrite 0x00000001", "fullread"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' not in captured.out
+    for address in range(100):
+        assert f'[Read] LBA {address:02} : 0x00000001' in captured.out
+
+@pytest.mark.skip
+@pytest.mark.parametrize("wrong_value", ["AA", "0xHELLO", "ox11", "0xaaaaaaaaaa", "-0xa", "1234", ";", " ", "0xA00000000000000001"])
+def test_shell_cmd_fullread_after_write_wrong_value(shell_app, wrong_value, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"fullwrite 0x00000001", f"fullwrite {wrong_value}", "fullread"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' in captured.out
+    for address in range(100):
+        assert f'[Read] LBA {address:02} : 0x00000001' in captured.out
 
 def test_shell_full_write_and_read_compare(shell_app, mocker: MockerFixture, capsys):
     # Arrange
-    shell_app._ssd_driver.run_ssd_write.side_effect = [WRITE_SUCCESS] * 100
-    shell_app._ssd_driver.run_ssd_read.side_effect = [READ_SUCCESS] * 100
-    shell_app._ssd_driver.get_ssd_output.return_value = "0x12345678"
+    cmd = [f"1_FullWriteAndReadCompare", f"1_"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
+
     # Act
-    shell_app.process_cmd("1_FullWriteAndReadCompare")
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
 
     # Assert
-    assert "Pass" in capsys.readouterr().out
-
-    assert shell_app._ssd_driver.run_ssd_write.call_count == 100
-    assert shell_app._ssd_driver.run_ssd_read.call_count == 100
-
+    assert 'INVALID COMMAND' not in captured.out
+    assert 'Pass' in captured.out
+    for address in range(100):
+        assert f'[Read] LBA {address:02} : 0x12345678' in captured.out
 
 def test_shell_partial_lba_write(shell_app, mocker: MockerFixture, capsys):
     # Arrange
-    shell_app._ssd_driver.run_ssd_write.side_effect = [WRITE_SUCCESS] * 150
-    shell_app._ssd_driver.run_ssd_read.side_effect = [READ_SUCCESS] * 150
-    shell_app._ssd_driver.get_ssd_output.return_value = "0x12345678"
+    cmd = [f"2_PartialLBAWrite", f"2_"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
     # Act
-    shell_app.process_cmd("2_PartialLBAWrite")
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
 
     # Assert
-    assert "Pass" in capsys.readouterr().out
+    assert 'INVALID COMMAND' not in captured.out
+    assert 'Pass' in captured.out
+    for address in range(5):
+        assert f'[Read] LBA {address:02} : 0x12345678' in captured.out #ToDo: check iterations are 30
 
-    assert shell_app._ssd_driver.run_ssd_write.call_count == 150
-    assert shell_app._ssd_driver.run_ssd_read.call_count == 150
-
-
+@pytest.mark.skip
 def test_shell_write_read_aging_with_real(shell_app, mocker: MockerFixture, capsys):
     # Arrange
-    shell_app = TestShellApp()
+    cmd = [f"3_WriteReadAging", f"3_"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
     # Act
-    shell_app.process_cmd("3_WriteReadAging")
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
 
     # Assert
-    assert "Pass" in capsys.readouterr().out
+    assert 'INVALID COMMAND' not in captured.out
+    assert 'Pass' in captured.out
+    #ToDo: check read value
 
+@pytest.mark.skip
+def test_shell_erase_write_aging(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    cmd = [f"4_EraseAndWriteAging", f"4_"]
+    cmd_len = len(cmd)
+    mocker.patch("builtins.input", side_effect=cmd)
 
-def test_shell_runner_with_testfile_correct_cmd(shell_app, mocker: MockerFixture, capsys):
+    # Act
+    ret = shell_app.run_shell(cmd_len)
+    captured = capsys.readouterr()
+
+    # Assert
+    assert 'INVALID COMMAND' not in captured.out
+    assert 'Pass' in captured.out
+    #ToDo: check read value
+
+@pytest.mark.skip
+@pytest.mark.parametrize("valid_cmd", ["1_", "2_", "3_", "4_", "1_FullWriteAndReadCompare",
+                                       "2_PartialLBAWrite", "3_WriteReadAging", "4_EraseAndWriteAging"])
+def test_shell_runner_with_testfile_valid_cmd(shell_app, valid_cmd, mocker: MockerFixture, capsys):
     # Arrange
     shell_app = TestShellApp()
     # 임시 파일 생성
     with tempfile.NamedTemporaryFile(mode='w+', delete=False, suffix='.txt') as temp_file:
-        temp_file.write("1_\n")
+        temp_file.write(f"{valid_cmd}\n")
         temp_file_path = temp_file.name
 
     # Act
@@ -353,15 +532,17 @@ def test_shell_runner_with_testfile_correct_cmd(shell_app, mocker: MockerFixture
         # 테스트 후 파일 삭제
         os.remove(temp_file_path)
     # Assert
-    assert "1_  ___  RUN...Pass" in capsys.readouterr().out
+    test_case = valid_cmd[:2]
+    assert f"{test_case}  ___  RUN...Pass" in capsys.readouterr().out
 
-
-def test_shell_runner_with_testfile_incorrect_cmd(shell_app, mocker: MockerFixture, capsys):
+@pytest.mark.parametrize("invalid_cmd", ["1_2", "he", " ", "None", "-",
+                                       "0_", "5_", "-1_"])
+def test_shell_runner_with_testfile_incorrect_cmd(shell_app, invalid_cmd, mocker: MockerFixture, capsys):
     # Arrange
     shell_app = TestShellApp()
     # 임시 파일 생성
     with tempfile.NamedTemporaryFile(mode='w+', delete=False, suffix='.txt') as temp_file:
-        temp_file.write("Hello\n")
+        temp_file.write(f"{invalid_cmd}\n")
         temp_file_path = temp_file.name
 
     # Act
@@ -374,113 +555,13 @@ def test_shell_runner_with_testfile_incorrect_cmd(shell_app, mocker: MockerFixtu
     # Assert
     assert "INVALID COMMAND" in capsys.readouterr().out
 
-
-def test_shell_runner_without_testfile(shell_app, mocker: MockerFixture, capsys):
+@pytest.mark.parametrize("invalid_file", [" ", f"{ROOT_DIR}\shell_scripts_.txt", f"shell_scripts.txt"])
+def test_shell_runner_with_invalid_testfile(shell_app, invalid_file, mocker: MockerFixture, capsys):
     # Arrange
     shell_app = TestShellApp()
 
     # Act
-    shell_app.run_runner()
+    shell_app.run_runner(invalid_file)
 
     # Assert
     assert "INVALID COMMAND" in capsys.readouterr().out
-
-
-def test_shell_runner_with_wrong_testfile(shell_app, mocker: MockerFixture, capsys):
-    # Arrange
-    shell_app = TestShellApp()
-
-    # Act
-    scipt_file = f"{ROOT_DIR}\shell_scripts_.txt"
-    shell_app.run_runner(scipt_file)
-
-    # Assert
-    assert "INVALID COMMAND" in capsys.readouterr().out
-
-
-@pytest.mark.parametrize("size", ["0.5", "-1.5", "0xa", "-0x1"])
-def test_shell_erase_invalid_size(shell_app, size):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
-
-    # Act
-    ret_fail = shell_app.erase(address="0", lba_size=size)
-
-    # Assert
-    assert ERASE_ERROR == ret_fail
-
-
-def test_shell_erase(shell_app):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
-
-    # Act
-    ret_pass = shell_app.erase(address="0", lba_size="1")
-
-    # Assert
-    assert ERASE_SUCCESS == ret_pass
-
-
-def test_shell_erase_resize(shell_app, mocker):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
-    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
-
-    # Act
-    ret_pass = shell_app.erase(address="0", lba_size="1000")
-
-    # Assert
-    assert ERASE_SUCCESS == ret_pass
-    shell_app._erase_in_chunks.assert_called_once_with(start_lba=0, size=100)
-
-
-def test_shell_erase_resize_minus(shell_app, mocker):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
-    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
-
-    # Act
-    ret_pass = shell_app.erase(address="30", lba_size="-20")
-
-    # Assert
-    assert ERASE_SUCCESS == ret_pass
-    shell_app._erase_in_chunks.assert_called_once_with(start_lba=11, size=20)
-
-
-def test_shell_erase_range(shell_app, mocker):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
-    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
-
-    # Act
-    ret_pass = shell_app.erase_range(start_lba="31", end_lba="60")
-
-    # Assert
-    assert ERASE_SUCCESS == ret_pass
-    shell_app._erase_in_chunks.assert_called_once_with(start_lba=31, size=30)
-
-
-def test_shell_erase_range_reverse(shell_app, mocker):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
-    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
-
-    # Act
-    ret_pass = shell_app.erase_range(start_lba="60", end_lba="31")
-
-    # Assert
-    assert ERASE_SUCCESS == ret_pass
-    shell_app._erase_in_chunks.assert_called_once_with(start_lba=31, size=30)
-
-
-@pytest.mark.parametrize("range", [("-1", "10"), ("1", "100"), ("1.5", "10"), ("10", "10.5")])
-def test_shell_erase_invalid_range(shell_app, mocker, range):
-    # Arrange
-    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
-    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
-
-    # Act
-    ret_pass = shell_app.erase_range(start_lba=range[0], end_lba=range[1])
-
-    # Assert
-    assert ERASE_ERROR == ret_pass
