+++ tests/test_shell_mock.py
@@ -1,0 +1,486 @@
+import tempfile
+
+import pytest
+from pytest_mock import MockerFixture
+from shell import *
+
+
+@pytest.fixture
+def shell_app(mocker):
+    ssd_driver_mock = mocker.Mock(spec=SSDDriver)
+    test_shell_app = TestShellApp(ssd_driver_mock)
+
+    return test_shell_app
+
+
+def test_shell_write_subprocess(shell_app):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
+
+    # Act
+    ret = shell_app.write(address="0", value="0x00000000")
+
+    # Assert
+    assert ret == WRITE_SUCCESS
+    shell_app._ssd_driver.run_ssd_write.assert_called_once_with(address="0", value="0x00000000")
+
+
+@pytest.mark.parametrize("wrong_address", ["-1", "100", "hello", "0.5", "-0.5", "123", ";"])
+def test_shell_write_wrong_address(shell_app, wrong_address):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
+
+    # Act
+    ret = shell_app.write(address=wrong_address, value="0x00000000")
+
+    # Assert
+    assert ret == WRITE_ERROR
+
+
+@pytest.mark.parametrize("wrong_value", ["AA", "0xHELLO", "ox11", "0xaaaaaaaaaa", "-0xa", "1234", ";"])
+def test_shell_write_wrong_value(shell_app, wrong_value):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
+
+    # Act
+    ret = shell_app.write(address="0", value=wrong_value)
+
+    # Assert
+    assert ret == WRITE_ERROR
+
+
+@pytest.mark.parametrize("valid_address", [str(x) for x in range(100)])
+def test_shell_write_test_valid_address(shell_app, valid_address):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
+
+    # Act
+    ret = shell_app.write(address=valid_address, value="0x00000001")
+
+    # Assert
+    assert ret == WRITE_SUCCESS
+    shell_app._ssd_driver.run_ssd_write.assert_called_once_with(address=valid_address, value="0x00000001")
+
+
+@pytest.mark.parametrize("valid_value", ["0xa", "0xab", "0xabc", "0xabcd", "0xabcde", "0xabcdef", "0xabcdeff"])
+def test_shell_write_valid_value(shell_app, valid_value):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
+
+    # Act
+    ret = shell_app.write(address="0", value=valid_value)
+
+    # Assert
+    assert ret == WRITE_SUCCESS
+
+
+def test_shell_read(shell_app, capsys):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_read.return_value = 0
+    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000000"
+
+    # Act
+    ret = shell_app.read(address="0")
+    captured = capsys.readouterr()
+
+    # Assert
+    assert ret == READ_SUCCESS
+    assert '[Read] LBA 00 : 0x00000000' in captured.out
+    shell_app._ssd_driver.run_ssd_read.assert_called_once_with(address="0")
+
+
+def test_shell_read_with_real(capsys):
+    # Arrange
+    shell_app = TestShellApp()
+
+    # Act
+    ret = shell_app.read(address="0")
+    captured = capsys.readouterr()
+
+    # Assert
+    assert ret == READ_SUCCESS
+
+
+def test_shell_read_after_write(shell_app, capsys):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.return_value = WRITE_SUCCESS
+    shell_app._ssd_driver.run_ssd_read.return_value = READ_SUCCESS
+    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000000"
+
+    # Act
+    ret_write = shell_app.write(address="0", value="0x00000000")
+    ret_read = shell_app.read(address="0")
+    captured = capsys.readouterr()
+
+    # Assert
+    assert ret_write == WRITE_SUCCESS
+    assert ret_read == READ_SUCCESS
+    assert '[Read] LBA 00 : 0x00000000' in captured.out
+
+    shell_app._ssd_driver.run_ssd_write.assert_called_once_with(address="0", value="0x00000000")
+    shell_app._ssd_driver.run_ssd_read.assert_called_once_with(address="0")
+
+
+def test_shell_read_after_read(shell_app, capsys):
+    # Arrange & Act
+    shell_app._ssd_driver.run_ssd_read.return_value = READ_SUCCESS
+    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000000"
+    ret0 = shell_app.read(address="0")
+
+    shell_app._ssd_driver.run_ssd_read.return_value = READ_SUCCESS
+    shell_app._ssd_driver.get_ssd_output.return_value = "0x00000010"
+    ret1 = shell_app.read(address="1")
+
+    captured = capsys.readouterr()
+
+    # Assert
+    assert ret0 == READ_SUCCESS
+    assert ret1 == READ_SUCCESS
+    assert '[Read] LBA 00 : 0x00000000' in captured.out
+    assert '[Read] LBA 01 : 0x00000010' in captured.out
+
+
+def test_shell_cmd_exit_success(shell_app, mocker: MockerFixture):
+    mocker.patch("builtins.input", side_effect=["exit"])
+    mock_method = mocker.patch("shell.TestShellApp.exit")
+    mock_method.return_value = 0
+    shell_app.run_shell(1)
+
+    mock_method.assert_called_once()
+
+
+def test_shell_cmd_read_success(shell_app, mocker: MockerFixture):
+    mocker.patch("builtins.input", side_effect=["read 1"])
+    mock_method = mocker.patch("shell.TestShellApp.read")
+    mock_method.return_value = 0
+    shell_app.run_shell(1)
+
+    mock_method.assert_called_once()
+
+
+def test_shell_cmd_write_success(shell_app, mocker: MockerFixture):
+    mocker.patch("builtins.input", side_effect=["write 0 0x00000001"])
+    mock_method = mocker.patch("shell.TestShellApp.write")
+    mock_method.return_value = 0
+    shell_app.run_shell(1)
+
+    mock_method.assert_called_once()
+
+
+def test_shell_cmd_help_success(shell_app, mocker: MockerFixture):
+    mocker.patch("builtins.input", side_effect=["help"])
+    mock_method = mocker.patch("shell.TestShellApp.help")
+    mock_method.return_value = 0
+    shell_app.run_shell(1)
+
+    mock_method.assert_called_once()
+
+
+def test_shell_cmd_fullread_success(shell_app, mocker: MockerFixture):
+    mocker.patch("builtins.input", side_effect=["fullread"])
+    mock_method = mocker.patch("shell.TestShellApp.full_read")
+    mock_method.return_value = 0
+    shell_app.run_shell(1)
+
+    mock_method.assert_called_once()
+
+
+def test_shell_cmd_fullwrite_success(shell_app, mocker: MockerFixture):
+    mocker.patch("builtins.input", side_effect=["fullwrite 0x00010000"])
+    mock_method = mocker.patch("shell.TestShellApp.full_write")
+    mock_method.return_value = 0
+    shell_app.run_shell(1)
+
+    mock_method.assert_called_once()
+
+
+@pytest.mark.parametrize("input", ["exi", "rea", "wri", "hel"])
+def test_shell_wrong_cmd(shell_app, mocker: MockerFixture, input, capsys):
+    mocker.patch("builtins.input", side_effect=[input])
+
+    shell_app.run_shell(1)
+    captured = capsys.readouterr()
+
+    assert 'INVALID COMMAND' in captured.out
+
+
+def test_shell_wrong_cmd_empty(shell_app, mocker: MockerFixture, capsys):
+    mocker.patch("builtins.input", side_effect=[""])
+
+    shell_app.run_shell(1)
+    captured = capsys.readouterr()
+
+    assert 'INVALID COMMAND' in captured.out
+
+
+@pytest.mark.parametrize("input", ["read 1 2", "write", "fullwrite", "exit 3"])
+def test_shell_wrong_cmd_args(shell_app, mocker: MockerFixture, input, capsys):
+    mocker.patch("builtins.input", side_effect=["read 1 2"])
+
+    shell_app.run_shell(1)
+    captured = capsys.readouterr()
+
+    assert 'INVALID COMMAND' in captured.out
+
+
+def test_shell_subprocess_cmd():
+    pass
+
+
+def test_shell_exit(shell_app, mocker: MockerFixture):
+    mocker.patch("builtins.input", side_effect=["exit"])
+    exit_mock = mocker.patch("shell.TestShellApp.exit", side_effect=SystemExit(0))
+
+    with pytest.raises(SystemExit) as e:
+        shell_app.run_shell(1)
+
+    exit_mock.assert_called_once()
+    assert e.value.code == 0
+
+
+def test_shell_help(capsys):
+    test_shell_app = TestShellApp()
+    test_shell_app.help()
+
+    captured = capsys.readouterr()
+    output_lines = captured.out.strip().splitlines()
+
+    expected_lines = [
+        "팀명: BestReviewer",
+        "팀장: 이장희 / 팀원: 김대용, 최도현, 박윤상, 최동희, 안효민, 김동훈",
+        "",
+        "사용 가능한 명령어:",
+        "  write <LBA> <Value>                : 특정 LBA에 값 저장",
+        "  read <LBA>                         : 특정 LBA 값 읽기",
+        "  erase <Start_LBA> <Size>           : Start_LBA부터 Size만큼 값 초기화",
+        "  erase_range <Start_LBA> <End_LBA>  : Start_LBA부터 End_LBA까지 값 초기화 ",
+        "  fullwrite <Value>                  : 전체 LBA에 동일 값 저장",
+        "  fullread                           : 전체 LBA 읽기 및 출력",
+        "  1_FullWriteAndReadCompare          : 전체 LBA 쓰기 및 비교",
+        "  2_PartialLBAWrite                  : LBA 0 ~ 4 쓰기 및 읽기 30회",
+        "  3_WriteReadAging                   : LBA 0, 99 랜덤 값 쓰기 및 읽기 200회",
+        "  4_EraseAndWriteAging               : LBA 짝수번호에 값을 두번 쓰고 및 size 3 만큼 지우기를 30회 반복함",
+        "  help                               : 도움말 출력",
+        "  exit                               : 종료"
+    ]
+
+    for line in expected_lines:
+        assert line in output_lines
+
+
+def test_shell_full_read(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_read.side_effect = [0] * 100
+    shell_app._ssd_driver.get_ssd_output.side_effect = ["0x00000000"] * 100
+
+    # Act
+    ret = shell_app.full_read()
+    captured = capsys.readouterr()
+    assert ret == READ_SUCCESS
+    assert shell_app._ssd_driver.run_ssd_read.call_count == 100
+
+
+def test_shell_full_write(shell_app, mocker: MockerFixture):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.side_effect = [WRITE_SUCCESS] * 100
+
+    # Act
+    ret = shell_app.full_write(value="0x12345678")
+
+    # Assert
+    assert ret == WRITE_SUCCESS
+    assert shell_app._ssd_driver.run_ssd_write.call_count == 100
+
+
+def test_shell_full_write_wrong_format():
+    pass
+
+
+def test_shell_full_write_and_read_compare(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.side_effect = [WRITE_SUCCESS] * 100
+    shell_app._ssd_driver.run_ssd_read.side_effect = [READ_SUCCESS] * 100
+    shell_app._ssd_driver.get_ssd_output.return_value = "0x12345678"
+    # Act
+    shell_app.process_cmd("1_FullWriteAndReadCompare")
+
+    # Assert
+    assert "Pass" in capsys.readouterr().out
+
+    assert shell_app._ssd_driver.run_ssd_write.call_count == 100
+    assert shell_app._ssd_driver.run_ssd_read.call_count == 100
+
+
+def test_shell_partial_lba_write(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_write.side_effect = [WRITE_SUCCESS] * 150
+    shell_app._ssd_driver.run_ssd_read.side_effect = [READ_SUCCESS] * 150
+    shell_app._ssd_driver.get_ssd_output.return_value = "0x12345678"
+
+    # Act
+    shell_app.process_cmd("2_PartialLBAWrite")
+
+    # Assert
+    assert "Pass" in capsys.readouterr().out
+
+    assert shell_app._ssd_driver.run_ssd_write.call_count == 150
+    assert shell_app._ssd_driver.run_ssd_read.call_count == 150
+
+
+def test_shell_write_read_aging_with_real(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app = TestShellApp()
+
+    # Act
+    shell_app.process_cmd("3_WriteReadAging")
+
+    # Assert
+    assert "Pass" in capsys.readouterr().out
+
+
+def test_shell_runner_with_testfile_correct_cmd(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app = TestShellApp()
+    # 임시 파일 생성
+    with tempfile.NamedTemporaryFile(mode='w+', delete=False, suffix='.txt') as temp_file:
+        temp_file.write("1_\n")
+        temp_file_path = temp_file.name
+
+    # Act
+    try:
+        shell_app.run_runner(temp_file_path)
+    finally:
+        # 테스트 후 파일 삭제
+        os.remove(temp_file_path)
+    # Assert
+    assert "1_  ___  RUN...Pass" in capsys.readouterr().out
+
+
+def test_shell_runner_with_testfile_incorrect_cmd(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app = TestShellApp()
+    # 임시 파일 생성
+    with tempfile.NamedTemporaryFile(mode='w+', delete=False, suffix='.txt') as temp_file:
+        temp_file.write("Hello\n")
+        temp_file_path = temp_file.name
+
+    # Act
+    try:
+        shell_app.run_runner(temp_file_path)
+    finally:
+        # 테스트 후 파일 삭제
+        os.remove(temp_file_path)
+
+    # Assert
+    assert "INVALID COMMAND" in capsys.readouterr().out
+
+
+def test_shell_runner_without_testfile(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app = TestShellApp()
+
+    # Act
+    shell_app.run_runner()
+
+    # Assert
+    assert "INVALID COMMAND" in capsys.readouterr().out
+
+
+def test_shell_runner_with_wrong_testfile(shell_app, mocker: MockerFixture, capsys):
+    # Arrange
+    shell_app = TestShellApp()
+
+    # Act
+    scipt_file = f"{ROOT_DIR}\shell_scripts_.txt"
+    shell_app.run_runner(scipt_file)
+
+    # Assert
+    assert "INVALID COMMAND" in capsys.readouterr().out
+
+
+@pytest.mark.parametrize("size", ["0.5", "-1.5", "0xa", "-0x1"])
+def test_shell_erase_invalid_size(shell_app, size):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
+
+    # Act
+    ret_fail = shell_app.erase(address="0", lba_size=size)
+
+    # Assert
+    assert ERASE_ERROR == ret_fail
+
+
+def test_shell_erase(shell_app):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
+
+    # Act
+    ret_pass = shell_app.erase(address="0", lba_size="1")
+
+    # Assert
+    assert ERASE_SUCCESS == ret_pass
+
+
+def test_shell_erase_resize(shell_app, mocker):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
+    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
+
+    # Act
+    ret_pass = shell_app.erase(address="0", lba_size="1000")
+
+    # Assert
+    assert ERASE_SUCCESS == ret_pass
+    shell_app._erase_in_chunks.assert_called_once_with(start_lba=0, size=100)
+
+
+def test_shell_erase_resize_minus(shell_app, mocker):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
+    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
+
+    # Act
+    ret_pass = shell_app.erase(address="30", lba_size="-20")
+
+    # Assert
+    assert ERASE_SUCCESS == ret_pass
+    shell_app._erase_in_chunks.assert_called_once_with(start_lba=11, size=20)
+
+
+def test_shell_erase_range(shell_app, mocker):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
+    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
+
+    # Act
+    ret_pass = shell_app.erase_range(start_lba="31", end_lba="60")
+
+    # Assert
+    assert ERASE_SUCCESS == ret_pass
+    shell_app._erase_in_chunks.assert_called_once_with(start_lba=31, size=30)
+
+
+def test_shell_erase_range_reverse(shell_app, mocker):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
+    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
+
+    # Act
+    ret_pass = shell_app.erase_range(start_lba="60", end_lba="31")
+
+    # Assert
+    assert ERASE_SUCCESS == ret_pass
+    shell_app._erase_in_chunks.assert_called_once_with(start_lba=31, size=30)
+
+
+@pytest.mark.parametrize("range", [("-1", "10"), ("1", "100"), ("1.5", "10"), ("10", "10.5")])
+def test_shell_erase_invalid_range(shell_app, mocker, range):
+    # Arrange
+    shell_app._ssd_driver.run_ssd_erase.return_value = ERASE_SUCCESS
+    mocker.patch.object(shell_app, "_erase_in_chunks", return_value=ERASE_SUCCESS, )
+
+    # Act
+    ret_pass = shell_app.erase_range(start_lba=range[0], end_lba=range[1])
+
+    # Assert
+    assert ERASE_ERROR == ret_pass
